+++ src/commands/index.ts
@@ -123,6 +123,7 @@ import { warn } from './moderation/warn/warn';
 import { listWarnings } from './moderation/warn/listWarnings';
 import { deleteWarn } from './moderation/warn/deleteWarn';
 import { atc } from './a32nx/atc';
+import { market } from './support/market';
 
 const commands: CommandDefinition[] = [
     ping,
@@ -248,6 +249,7 @@ const commands: CommandDefinition[] = [
     listWarnings,
     deleteWarn,
     atc,
+    market,
 ];
 
 const commandsObject: { [k: string]: CommandDefinition } = {};