+++ .github/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Update <small>_ July 2022</small>
 
+- feat: marketplace removal command (11/07/2022)
 - feat: default ATC command (10/07/2022)
 - fix: remove flypados thread from .exp (07/07/2022)